\name{genecards}
\alias{genecards}
\title{GeneCards database url}
\description{the GeneCards database url at the Weizmann Institute}
\usage{genecards}
\value{a character string containing all but the Affymetrix
 gene identifier giving a valid search querey to the GeneCards
 database. Used in TopGenes. The user can supply her/his own
 using this one as a template.}
\author{Grant Izmirlian \email{izmirlian@nih.gov}}
\seealso{\code{\link{TopGenes}}}

\keyword{utilities}
